% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cw_as_standalone.R
\name{cw_as_standalone}
\alias{cw_as_standalone}
\title{Turn crosswords into standalone}
\usage{
cw_as_standalone(crossword)
}
\arguments{
\item{crossword}{a crossword}
}
\value{
Either write to a file or returns the result as character vector to be
handled by the user.
}
\description{
Turn crosswords into standalone
}
